Delete GRBL-code info
